Auth and Application class implemented
