+++ app/src/main/java/com/example/bogdan/dropboxphoto/infrastructure/DbxApplication.java
@@ -1,0 +1,9 @@
+package com.example.bogdan.dropboxphoto.infrastructure;
+
+import android.app.Application;
+
+public class DbxApplication extends Application {
+    private static final String APP_KEY = "e7r6jtptl6t3rz9";
+    private static final String APP_SECRET = "qqfvu5wtkqft9uz";
+
+}
